fix: validate mandatory fields and hsn codes for e-Invoice generation
(cherry picked from commit c28a2ebaeb161693951ef61ea2e4280d2815952f)
